% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-region-index.R
\name{append_region_index}
\alias{append_region_index}
\title{Extract Region Index from Spatial Object and Partition Training and Testing Data}
\usage{
append_region_index(train_data, polys)
}
\arguments{
\item{train_data}{A data frame containing training data.}

\item{States}{Shapefile object containing state boundaries and regions.}

\item{train_end_date}{A Date object till which we want to keep the data for training purposes.}
}
\value{
Returns the modified train_data dataframe with new columns: day, trn_tst, and Region.
}
\description{
This function extracts region data for each location from a given shapefile. It also adds additional columns for day of week and training or testing data flag.
}
\examples{
append_region_index(train_df, states_sf, as.Date("2020-12-31"))

}
