% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-date-string.R
\name{cast_date_string}
\alias{cast_date_string}
\title{Convert Date String to Date Object}
\usage{
cast_date_string(x, null_replacement_value)
}
\arguments{
\item{x}{Date string in YYYY-MM-DD format}

\item{null_replacement_value}{Value to return if input is null}
}
\value{
A date object representing the input string
}
\description{
Takes a date string in YYYY-MM-DD format and returns a date object, or a null replacement value if input is null.
}
\examples{
cast_date_string("2022-03-02", "No date")
cast_date_string(NULL, "No date")
}
