% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-shapefile.R
\name{download_boundaries}
\alias{download_boundaries}
\title{Downloads and reads US jurisdiction boundary polygons}
\usage{
download_boundaries(output_dir, unit = "state", proj = "EPSG:5070")
}
\arguments{
\item{output_dir}{character string specifying the path/to/output/directory where the extracted shapefiles will be stored}

\item{unit}{character string specifying the level of geographic unit for which shapefile is needed. Defaults to "state".}

\item{proj}{character string specifying the desired coordinate reference system (CRS). Defaults to "EPSG:5070".}
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
Downloads and reads US jurisdiction boundary polygons
}
\examples{
download_boundaries(output_dir = "C:/MyFiles", unit = "county", proj = "EPSG:4326")

}
