% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-shapefile.R
\name{download_shapefile}
\alias{download_shapefile}
\title{Downloads and reads US state boundaries shapefiles into an `sf` object}
\usage{
download_shapefile(output_dir)
}
\arguments{
\item{output_dir}{The output directory where the shapefiles will be saved and read from.}
}
\value{
An `sf` object containing the US state boundaries shapefiles.
}
\description{
This function downloads the US state boundaries shapefiles from a specific URL and saves them to the specified
output directory. The downloaded zip file is then unzipped and the resulting shapefiles are read into an `sf`
object using the 'st_read' function from the 'sf' package.
}
\examples{
States <- download_shapefile("./polygons")
}
