% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-api.R
\name{get_api}
\alias{get_api}
\title{Get API key from secrets.yaml file}
\usage{
get_api(api_name)
}
\arguments{
\item{api_name}{A character string representing the name of the API key to be retrieved from secrets.yaml file.}
}
\value{
A character vector containing the value of the requested API key.
}
\description{
This function retrieves an API key by reading a secrets.yaml file stored in the current working directory. The specified 'api_name' is used to retrieve the corresponding API key.
}
\examples{
#Read API token for Twitter API from secrets.yaml
get_api("Twitter_API")

#Read API token for OpenWeather API from secrets.yaml
get_api("OpenWeather_API")

}
