% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-hub-forecasts.R
\name{get_hub_forecasts}
\alias{get_hub_forecasts}
\title{Obtain COVID-19 Observation Data}
\usage{
get_hub_forecasts(
  source = c("covidHubUtils", "cache", "test"),
  models = NULL,
  write_copy = TRUE
)
}
\arguments{
\item{source}{A character string indicating the source of the data to be obtained. Allowed values are "covidcast" (to fetch data using covidcast), "cache" (to read data from local cache) and "test" (to load test data).}

\item{write_copy}{A logical value indicating whether the data obtained should be written to a file or not. The default value is TRUE.}

\item{start_date}{A character string indicating the start date of the period for which the data is needed. The date should be in the format "YYYY-MM-DD".}

\item{end_date}{A character string indicating the end date of the period for which the data is needed. The date should be in the format "YYYY-MM-DD".}
}
\value{
Returns a data frame containing COVID-19 observation data based on the chosen source.
}
\description{
This function obtains the COVID-19 observation data based on the chosen source.
}
\details{
If source = "covidcast", the function fetches the COVID-19 observation data using the covidcast package. If source = "cache", the function reads the COVID-19 observation data from the local cache. If source = "test", the function loads test data for the period 2021-06-28 through 2021-09-20. The function also loads state crosswalk data and performs some date checks and data filtering based on the selected source.

If write_copy = TRUE, the data obtained is written to a .csv file in a newly created directory named "observations" within the analysis directory.
}
