% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-model-name.R
\name{get_model_name}
\alias{get_model_name}
\title{Get the model name from a filename}
\usage{
get_model_name(filename)
}
\arguments{
\item{filename}{A character string representing the filename.}
}
\value{
A character string representing the model name.
}
\description{
This function takes in a filename in .csv format and returns the model name. The function removes the ".csv" extension, finds the last occurrence of "-" in the filename, and removes all text preceding this last dash (including the dash itself).
}
\examples{
get_model_name("model-1.csv")
[1] "1"
get_model_name("results-model-2.csv")
[1] "2"
}
