% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-neighbors.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get Neighbors of Polygons}
\usage{
get_neighbors(polys, distance = "centroid", connect = TRUE)
}
\arguments{
\item{polys}{A SpatialPolygons object containing the polygons for which to
calculate neighbourhoods.}

\item{distance}{The method by which to calculate distances between polygons.
Set to "centroid" (default) to use the distance between centroids of polygons,
or "polygon" to measure distances between the polygons themselves.}

\item{connect}{Logical. Whether to create a neighborhood list based on
polygon adjacency (TRUE) or polygon connectivity (FALSE).}
}
\value{
A 'neighbours' object created using the 'poly2nb' function from the
spdep package. The neighbours object is a list of integer vectors, with one
vector per spatial unit (polygon), describing indices into the \code{polys}
object to define neighbouring spatial units.
}
\description{
This function takes a set of polygons and calculates neighbors based on
either polygon adjacency or polygon distance from centroid.
}
\examples{
library(spdep)
data(nc.sids)
neighbors <- get_neighbors(polygons = nc.sids, distance="polygon", connect=FALSE)

}
