% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-poverty-pop.R
\name{getPovertyPop}
\alias{getPovertyPop}
\title{Returns a merged dataframe with poverty and population data by state for specified age range.}
\usage{
getPovertyPop(key, vars_pov, vars_pop, filt_age)
}
\arguments{
\item{key}{A character string containing valid value of user's Census API key.}

\item{vars_pov}{A character vector that specifies the variables to be retrieved for poverty data.}

\item{vars_pop}{A character vector that specifies the variables to be retrieved for population data.}

\item{filt_age}{An integer vector of length two specifying the minimum and maximum age group to filter population data.}
}
\value{
A merged dataframe with poverty and population count for each state.
}
\description{
This function retrieves poverty data and population data grouped by age from the Census API, filters the
data and processes it to return a merged dataframe with relevant information.
}
\examples{
data <- getPovertyPop("my_census_key", c("SAEMHI_PT"), c("POP"), c(12, 18))
}
