% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-inla.R
\name{inla_dlm}
\alias{inla_dlm}
\title{Dynamic Linear Model using INLA}
\usage{
inla_dlm(ts_data, forecast_horizon)
}
\arguments{
\item{ts_data}{Time-series data as numeric vector}

\item{forecast_horizon}{The number of periods for which forecasts are required}
}
\value{
The function returns a list with two elements:
\itemize{
\item \code{obs_fitted}: A numeric vector representing the fitted values over the observation period including a 7-day lag (EpiEstim).
\item \code{pred_trend}: A numeric vector representing the predicted values for the forecast horizon.
}
}
\description{
This function fits a dynamic linear model (DLM) to the input time series data, and produces forecasts for the specified number of periods using the INLA method.
}
\examples{
inla_dlm(c(10,12,16,18,22,26,30,34,40),5)

}
