% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-location.R
\name{plot_location}
\alias{plot_location}
\title{Plot Location}
\usage{
plot_location(plot_path = plot_paths, model = NULL, loc = NULL)
}
\arguments{
\item{plot_path}{A character vector containing the paths of the CSV files with the forecasted data.}

\item{model}{A character vector specifying the type of model.}

\item{loc}{A character vector speciyfing the location to be visualized. If this parameter is NULL, the function will generate reports for all available locations. Default value is NULL.}
}
\value{
This function returns a ggplot object containing the visualization of hospital incidence in the specified location(s) with forecasted data.
}
\description{
The function \code{plot_location} is used to create a report on hospital incidence in the provided location(s) with forecasted data. The report can include all locations or specific locations as per user input.
}
