% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-neighbors.R
\name{plot_neighbors}
\alias{plot_neighbors}
\title{Plots neighborhood adjacency}
\usage{
plot_neighbors(poly_obj, nb_obj)
}
\arguments{
\item{poly_obj}{A SpatialPolygonsDataFrame representing the polygons.}

\item{nb_obj}{A nb object generated using the 'spdep' package representing the neighborhood relationship information.}
}
\value{
A ggplot object with the plotted polygons and their neighboring relationships.
}
\description{
This function takes in a polygon object and its corresponding neighborhood object,
and plots the polygons and their neighbors on a map.
}
\details{
The function first extracts the coordinates from \code{poly_obj} and creates a data frame named \code{coords}
with columns "long" and "lat". Then, it generates a set of lines representing the neighboring relationships between polygons
using the \code{nb2lines} function. It then plots the polygons and their neighbors on a map using ggplot2.
}
\examples{
library(sp)
library(spdep)
library(ggplot2)
library(sf)

data(columbus)
nb_obj <- dnearneigh(coordinates(columbus), 0, 10000)
poly_obj <- columbus

plot_neighbors(poly_obj, nb_obj)

}
