% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-wis-lines.R
\name{plot_WIS_lines}
\alias{plot_WIS_lines}
\title{Plot WIS Lines Function}
\usage{
plot_WIS_lines(
  scores_df,
  by = c("tile", "date"),
  range = c("abs", "scaled"),
  scale_model = NULL,
  limit = NULL
)
}
\arguments{
\item{scores_df}{A dataframe containing the WIS scores, date, model, and location_name columns.}

\item{by}{A character vector specifying how to plot the data. Valid options are "tile" (heatmap) and "date" (line chart).}

\item{range}{A character vector specifying the type of scaling to apply to the WIS scores. Valid options are "abs" (absolute scaling) and "scaled" (scaling based on another model's WIS scores).}

\item{scale_model}{A character vector indicating the model to use for scaling when range = "scaled".}

\item{limit}{A numeric vector specifying the maximum value for the WIS score.}
}
\value{
This function returns a plot of the WIS scores according to the specified parameters.
}
\description{
This function plots a line chart or heatmap of the Weighted Interval Score (WIS) for different models and/or locations over time.
}
