% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt-projection.R
\name{Rt_projection}
\alias{Rt_projection}
\title{Project Rt values using ARIMA or DLM method}
\usage{
Rt_projection(
  train_data,
  forecast_horiz_start,
  forecast_horiz_end,
  mean_si = 5.7,
  std_si = 2,
  method = c("arima", "dlm")
)
}
\arguments{
\item{train_data}{A dataframe with columns for date, location_name, and value, where value is a numeric vector representing the number of new cases or incidences for that day and location.}

\item{forecast_horiz_start}{A date string representing the start of the time period to project Rt values for.}

\item{forecast_horiz_end}{A date string representing the end of the time period to project Rt values for.}

\item{mean_si}{A numeric value representing the mean serial interval for the disease being modeled. Default is 5.7.}

\item{std_si}{A numeric value representing the standard deviation of the serial interval for the disease being modeled. Default is 2.}

\item{method}{A character vector specifying the method to use for projecting Rt values. Valid options are "arima" or "dlm". Default is "arima".}
}
\value{
A dataframe with the same columns as `train_data`, plus an additional column `Rt` representing the projected reproductive rate for each day and location.
}
\description{
This function projects Rt (reproductive rate) values for a given time period and location using either an ARIMA or DLM model.
}
\examples{
# Load example data
data(covid_cases)

# Project Rt values using default parameters and ARIMA method
Rt_df <- Rt_projection(train_data = covid_cases,
                        forecast_horiz_start = "2020-04-12",
                        forecast_horiz_end = "2020-04-26")

# Project Rt values using DLM method
Rt_df <- Rt_projection(train_data = covid_cases,
                        forecast_horiz_start = "2020-04-12",
                        forecast_horiz_end = "2020-04-26",
                        method = "dlm")

}
