% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model-list.R
\name{run_model_list}
\alias{run_model_list}
\title{Run a series of INLA models}
\usage{
run_model_list(
  formulas.list,
  dataStack,
  likelihood = "gaussian",
  config = FALSE,
  verbose = FALSE,
  archive = TRUE
)
}
\arguments{
\item{formulas.list}{A named list of model formulas to be run. Names should match the desired output names.}

\item{dataStack}{Stacked data object as created by \code{inla.stack.data()} function from the INLA package.}

\item{likelihood}{A character vector specifying which likelihood family to use for each model. If a length 1 character string is provided, it will be used for all models. If a character vector of length n_models is provided, it will be applied in order.}

\item{config}{A logical value indicating whether or not to retain GMRF representation for sampling.}

\item{verbose}{A logical value indicating whether or not to print details to screen while running.}

\item{archive}{A logical value indicating whether or not to save the output in an archive file. Default is TRUE.}
}
\value{
models_out A list of model outputs with names corresponding to the input formula names.
}
\description{
This function runs a list of INLA models specified in \code{formulas.list} on data stacked in \code{dataStack}. The function returns a list of model outputs with names matching the input formula names.
}
