% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-time-index.R
\name{time_index}
\alias{time_index}
\title{Generate a temporal index for a given set of dates}
\usage{
time_index(x, y)
}
\arguments{
\item{x}{A date vector, typically a column in a dataframe indicating date of observed case or incidence.}

\item{y}{A vector of dates to search for matches, typically a customized series with specified interval length.}
}
\value{
A vector of dates from \code{y} that are nearest matches for dates given in \code{x}
}
\description{
This function returns the date from \code{y} that is the closest to each date in \code{x}.
}
