% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-forecast-template.R
\name{create_forecast_template}
\alias{create_forecast_template}
\title{This function creates a forecast template using the input training data and specific dates for forecasting.}
\usage{
create_forecast_template(train_data, forecast_horiz_start, forecast_horiz_end)
}
\arguments{
\item{train_data}{A data frame containing the training data with columns for location, date, and value}

\item{forecast_horiz_start}{The starting date for the forecast horizon in yyyy-mm-dd format as a character string}

\item{forecast_horiz_end}{The ending date for the forecast horizon in yyyy-mm-dd format as a character string}
}
\value{
A data frame containing the forecast template with columns for location, date, value, signal, day, trn_tst, and region_table
}
\description{
This function creates a forecast template using the input training data and specific dates for forecasting.
}
