% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-shapefile.R
\name{download_boundaries}
\alias{download_boundaries}
\title{Download US Boundaries}
\usage{
download_boundaries(unit = "state", proj = "EPSG:5070")
}
\arguments{
\item{unit}{A character string indicating whether to download state or county borders, default is "state".}

\item{proj}{Projection for the shapefiles, defaults to "EPSG:5070".}
}
\value{
An object of class 'SpatialPolygonsDataFrame' which contains the state or county border polygons and their associated attributes.
}
\description{
Downloads and extracts shapefiles for either US states or counties. It creates a directory in the specified location if one does not exist already. The function downloads the shapefile from the specified URL, unzips it, reads it into an object of class 'sf', changes its projection to that of "EPSG:5070", fixes topology, matches state name with observation data and assigns unique identifiers.
}
\examples{
\donotrun{
#Download state boundaries to a directory called 'States':
download_boundaries("./States")

#Download county boundaries to a directory called 'Counties':
download_boundaries("./Counties", unit = "county")
}

}
