% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-covid19-obs.R
\name{get_covid19_obs}
\alias{get_covid19_obs}
\title{Obtain COVID-19 observed data based on the chosen source}
\usage{
get_covid19_obs(
  source = c("covidcast", "cache", "test"),
  start_date,
  end_date,
  write_copy = TRUE
)
}
\arguments{
\item{source}{A character string indicating the source of data to be obtained. Possible values are "covidcast" for obtaining data from the covidcast API, "cache" for obtaining data from local cache and "test" for obtaining test data.}

\item{start_date}{A character string in the format "YYYY-MM-DD" indicating the start date for the data to be obtained.}

\item{end_date}{A character string in the format "YYYY-MM-DD" indicating the end date for the data to be obtained.}
}
\value{
Data frame containing COVID-19 observed data.
}
\description{
This function obtains COVID-19 observed data from either the covidcast API, local cache or test data and returns them as a data frame.
}
